requirements.txt
alabaster==1.0.0
anyio==4.11.0
asgiref==3.10.0
av==16.0.1
babel==2.17.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.1
Cython==3.2.1
Django==5.2.8
djangorestframework==3.16.1
docutils==0.21.2
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
indic-nlp-library-itt==0.1.1
indictranstoolkit==1.1.1
Jinja2==3.1.6
joblib==1.5.2
langdetect==1.0.9
lxml==6.0.2
MarkupSafe==3.0.2
Morfessor==2.0.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
pillow==12.0.0
portalocker==3.2.0
protobuf==6.33.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
roman-numerals-py==3.1.0
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
transformers==4.57.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
